api/pyproject.toml
[tool.poetry]
name = "ai4mde-studio"
version = "1.0.0"
description = "AI-assisted Model Driven Engineering"
authors = ["Max Boone <contact@maxboone.com>"]
license = "GPLv3"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
django = "5.0.2"
psycopg2-binary = "^2.9.9"
django-ninja = "^1.1.0"
pyjwt = "^2.8.0"
django-cors-headers = "^4.3.1"
daphne = "^4.0.0"
channels = "^4.0.0"
sentry-sdk = "^1.40.4"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ruff = "^0.1.13"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "model.settings"
